Update Example to OF 0.9.0
