+++ example/src/main.cpp
@@ -1,11 +1,13 @@
+#include "ofMain.h"
 #include "testApp.h"
-#include "ofAppGlutWindow.h"
 
 //--------------------------------------------------------------
 int main()
 {
-	ofAppGlutWindow window; // create a window
-	// set width, height, mode (OF_WINDOW or OF_FULLSCREEN)
-	ofSetupOpenGL(&window, 1024, 768, OF_WINDOW);
-	ofRunApp(new testApp()); // start the app
+    ofSetupOpenGL(1024,768, OF_WINDOW);			// <-------- setup the GL context
+    
+    // this kicks off the running of my app
+    // can be OF_WINDOW or OF_FULLSCREEN
+    // pass in width and height too:
+    ofRunApp( new testApp());
 }
