+++ example/src/testApp.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "ofMain.h"
+#include "MyRect.h"
 
 class testApp : public ofBaseApp
 {
@@ -8,14 +9,11 @@ public:
 	void setup();
 	void update();
 	void draw();
+    void keyPressed(int key);
 
-	void keyPressed(int key);
-	void keyReleased(int key);
-	void mouseMoved(int x, int y);
-	void mouseDragged(int x, int y, int button);
-	void mousePressed(int x, int y, int button);
-	void mouseReleased(int x, int y, int button);
-	void windowResized(int w, int h);
-	void dragEvent(ofDragInfo dragInfo);
-	void gotMessage(ofMessage msg);
+    void addRects();
+    
+    vector<MyRect*> rects;
+    ofxInteractivePrimitives::RootNode *root;
+    
 };
